fix: resolve compilation errors and create database migrations
**Fixed Issues:**
1. tower-http dependency: Updated to use compression-gzip and compression-br features (V0.5)
2. Created missing database migrations (migrations/20250103000000_init_schema.sql and .down.sql)
   - 8 tables with proper constraints, indexes, and foreign keys
   - PostgreSQL-specific features (CITEXT, JSONB, UUID, ENUMs)
   - Complete UP and DOWN migrations
3. Added axum and hmac dependencies to axontask-shared
4. Fixed Hmac import in webhook.rs (use hmac crate instead of sha2::Hmac)
5. Fixed lifetime issues in auth middleware IntoResponse impl
6. Fixed AuthContext.user_id to be Option<Uuid> (API keys don't have users)
7. Updated require_role to accept Option<Uuid>
8. Fixed type mismatches in stream_reader (k.as_ref() type annotations)
9. Fixed pool stats type mismatch (idle as u32)
10. Added ApiKey::is_expired() method
11. Added MembershipRole::has_permission() method
12. Fixed module doc comments (///! -> //!)

**Status:**
- ✅ axontask-shared: compiles successfully with warnings only
- ⏳ axontask-api: has compilation errors (Extension, CreateTenant fields, Display traits)
- ⏳ axontask-worker: has compilation errors (to be fixed)

+++ axontask-shared/src/db/pool.rs
@@ -234,8 +234,8 @@ pub fn get_pool_stats(pool: &PgPool) -> PoolStats {
     let idle = pool.num_idle();
 
     PoolStats {
-        active_connections: size.saturating_sub(idle) as usize,
-        idle_connections: idle as usize,
+        active_connections: (size.saturating_sub(idle as u32)) as usize,
+        idle_connections: idle,
         total_connections: size as usize,
     }
 }
